[HAL][RTC] Add explicit cast into RTC_ByteToBcd2() to avoid warning about implicit narrowing conversion

+++ Src/stm32wbxx_hal_rtc.c
@@ -1914,7 +1914,7 @@ uint8_t RTC_ByteToBcd2(uint8_t number)
   while (bcdlow >= 10U)
   {
     bcdhigh++;
-    bcdlow -= 10U;
+    bcdlow -= (uint8_t)10;
   }
 
   return ((uint8_t)(bcdhigh << 4U) | bcdlow);
